backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
requests
groq
tavily-python
pymongo
